all chopped up
