+++ src/main/java/com/example/movingaverage/Global.java
@@ -1,5 +1,6 @@
 package com.example.movingaverage;
 
+import com.example.movingaverage.DAO.MongoCRUD;
 import com.google.api.client.http.HttpRequestFactory;
 import com.google.api.client.http.javanet.NetHttpTransport;
 
@@ -7,6 +8,7 @@ import java.time.LocalDateTime;
 
 public class Global {
     // if we abstract the markets now we will affect the DAO layer as well as the price object
+    public static MongoCRUD mongoCRUD = MongoCRUD.getInstance();
     public static LocalDateTime start = null;
     public static String orderTimeInForce = "FILL_OR_KILL";
     public static double quant;
