+++ src/test/java/com/example/movingaverage/MainTest.java
@@ -1,0 +1,20 @@
+package com.example.movingaverage;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayDeque;
+import java.util.Deque;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class MainTest {
+    Deque<String> commands = new ArrayDeque<>(10);
+
+    @Test
+    void bootTest() {
+        //Add The commands to the deque
+        commands.add("boot");
+
+    }
+
+}
